+++ Sources/Sentry/include/SentryWatchdogTerminationScopeObserver.h
@@ -5,9 +5,9 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
-/// This scope observer is used by the Watchdog Terminations integration to write breadcrumbs to
+/// This scope observer is used by the Watchdog Termination integration to write breadcrumbs to
 /// disk. The overhead is ~0.015 seconds for 1000 breadcrumbs.
-@interface SentryWatchdogTerminationsScopeObserver : NSObject <SentryScopeObserver>
+@interface SentryWatchdogTerminationScopeObserver : NSObject <SentryScopeObserver>
 SENTRY_NO_INIT
 
 - (instancetype)initWithMaxBreadcrumbs:(NSInteger)maxBreadcrumbs